+++ resources/js/api/user.js
